+++ server/src-lib/Hasura/Backends/Postgres/SQL/Types.hs
@@ -39,6 +39,7 @@ module Hasura.Backends.Postgres.SQL.Types
     qualifiedObjectToName,
     PGScalarType (..),
     textToPGScalarType,
+    pgScalarTranslations,
     pgScalarTypeToText,
     PGTypeKind (..),
     QualifiedPGType (..),